feat(decision-lab): 加入五軸 Probe sizing
以 weakest-link ceiling、paper/live 各自的 book、factor 與 execution liquidity 約束產生可重現部位範圍。政策與 rubric 版本化，跨掛牌交易資料缺失時只封鎖 live lane。

+++ decision_lab/adapters/market.py
@@ -34,7 +34,7 @@ def normalize_market_snapshot(
     expected_ticker: str,
     expected_currency: str,
     evaluation_at: str,
-    max_age_hours: int = 36,
+    max_age_hours: float = 36,
 ) -> dict[str, Any]:
     upstream = payload.get("status")
     if upstream in {"missing", "unavailable"}:
@@ -85,7 +85,7 @@ def normalize_fx_snapshot(
     *,
     expected_pair: str,
     evaluation_at: str,
-    max_age_hours: int = 36,
+    max_age_hours: float = 36,
 ) -> dict[str, Any]:
     upstream = payload.get("status")
     if upstream in {"missing", "unavailable"}:
